Update homebrew_tap to support custom tap URL via optional 'url' option
This allows doing 'brew tap <url>' where the URL is not assumed to be on GitHub,
and the protocol doesn't have to be HTTP. Any location and protocol that git can
handle is fine.

While at it, allow proper `list` type support for 'name' option and update
module documentation for option aliases.